requirements.txt
flask==2.3.3
openai==1.23.2
numpy==1.26.4
playwright==1.44.0
python-dotenv==1.0.1
langdetect==1.0.9
